==> src/10-plot-hiscos.qmd <==
---
title: "Plot hiscos"
format: gfm
---



```{r}
library(tidyverse)
theme_set(theme_light())

df = readxl::read_excel("umap_occupations.xlsx") %>% janitor::clean_names()

df_codes = readxl::read_excel('data/occupations/3-digit-hisco.xlsx') %>% janitor::clean_names()

df = df %>% 
mutate(name = str_extract(name, ".* //"),
name = str_squish(str_remove(name, "//")))

df = df %>% inner_join(df_codes, by = "name") %>% 
mutate(first_digit = substr(number, 1, 1)) %>% 
mutate(first_digit = as.numeric(first_digit))

library(tibble)

hisco_groups <- tibble(
  first_digit = c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9),
  group_description = c(
    "Professional, technical and related workers",
    "Professional, technical and related workers",
    "Administrative and managerial workers",
    "Clerical and related workers",
    "Sales Workers",
    "Service workers",
    "Agricultural, animal husbandry and forestry workers, fishermen and hunters",
    "Production and related workers, transport equipment operators and labourers",
    "Production and related workers, transport equipment operators and labourers",
    "Production and related workers, transport equipment operators and labourers"
  )
)

df = df %>% 
inner_join(hisco_groups)

df %>% select(group_description)

library(ggrepel)

set.seed(2024)

df %>% 
slice_sample(n = 20) %>% 
ggplot(aes(x = umap_component_1, umap_component_2, colour = group_description)) +
geom_point() +
geom_text_repel(aes(label = name), show.legend = F) +
theme(legend.position = "bottom") + 
guides(colour=guide_legend(nrow=7,byrow=TRUE)) +
labs(x = "UMAP component 1", y =  "UMAP component 2", colour = "HISCO major group")

ggsave("figures/hisco_cluster.png", width = 20, height = 25, units = "cm", dpi = 300)
```
